+++ src/LCB_Clone.Api/Services/Interfaces/ILegislatorStringsServices.cs
@@ -5,8 +5,8 @@ namespace LCB_Clone.Api.Services.Interfaces;
 public interface ILegislatorStringsServices
 {
 	Task<List<LegislatorStringsResponseDto>> GetAll();
-	Task<LegislatorStringsResponseDto?> GetOne(int id);
+	Task<LegislatorStringsResponseDto?> GetOne(ulong id);
 	Task<LegislatorStringsResponseDto> Create(LegislatorStringsCreateDto dto);
 	Task<LegislatorStringsResponseDto?> Update(LegislatorStringsUpdateDto dto);
-	Task<bool> Delete(int id);
+	Task<bool> Delete(ulong id);
 }
