+++ src/LCB_Clone.Api/Services/LegislatorStringsServices.cs
@@ -47,7 +47,7 @@ public class LegislatorStringsServices(AppDbContext db) : ILegislatorStringsServ
 		return legislatorStrings;
 	}
 
-	public async Task<LegislatorStringsResponseDto?> GetOne(int id)
+	public async Task<LegislatorStringsResponseDto?> GetOne(ulong id)
 	{
 		LegislatorStringsResponseDto? legislatorString = await _db.LegislatorStrings
 			.AsNoTracking()
@@ -106,7 +106,7 @@ public class LegislatorStringsServices(AppDbContext db) : ILegislatorStringsServ
 		return null;
 	}
 
-	public async Task<bool> Delete(int id)
+	public async Task<bool> Delete(ulong id)
 	{
 		return await _db.LegislatorStrings.Where(ls => ls.Id == id).ExecuteDeleteAsync() > 0;
 	}
