+++ src/LCB_Clone.Shared/Dtos/LegislatorStrings/LegislatorStringsResponseDto.cs
@@ -4,9 +4,9 @@ using LCB_Clone.Shared.Enums.Legislators;
 namespace LCB_Clone.Shared.Dtos.LegislatorStrings;
 
 public record LegislatorStringsResponseDto(
-	int Id,
+	ulong Id,
 	string Text,
 	LegislatorStringType Type,
-	int LegislatorId,
+	ulong LegislatorId,
 	LegislatorResponseDto? Legislator
 );
